removed recoil off guns as the game is going in a more fast pased style, leaving recoil system in the game incase it is needed later

+++ 2D top down/Character.cpp
@@ -166,7 +166,7 @@ void Player::FireWeapon()
 					{
 						SDL_Point tempPoint = GetMapCoordFromCursor();
 
-						Vector2D aimLoc = Vector2D(tempPoint.x, tempPoint.y);
+						Vector2D aimLoc = Vector2D(static_cast <float> (tempPoint.x), static_cast <float>  (tempPoint.y) );
 
 						// Calculate deviation/recoil etc.
 						calcDeviation = this->weapon[this->selectedWeapon]->deviation + currentRecoil;
